♻️ #238: MyInfo 데이터 로직, MainRepository에서 MyInfoRepository로 마이그레이션
