+++ StreetDrop/StreetDrop/Data/Repositories/Main/DefaultMainRepository.swift
@@ -20,14 +20,6 @@ final class DefaultMainRepository: MainRepository {
 }
 
 extension DefaultMainRepository {
-    func fetchMyInfo() -> Single<MyInfo> {
-        return networkManager.getMyInfo()
-            .map({ data in
-                let dto = try JSONDecoder().decode(MyInfoResponseDTO.self, from: data)
-                return dto.toEntity()
-            })
-    }
-    
     func fetchPoi(lat: Double, lon: Double, distacne: Double) -> Single<Pois> {
         networkManager.getPoi(latitude: lat, longitude: lon, distance: distacne)
             .map({ data in
@@ -50,9 +42,5 @@ extension DefaultMainRepository {
                 let dto = try JSONDecoder().decode(MusicWithinAreaResponseDTO.self, from: data)
                 return dto.toEntity()
             })
-    }
-    
-    func saveMyInfo(_ myInfo: MyInfo) -> RxSwift.Single<Void> {
-        return myInfoStorage.saveMyInfo(myInfo: myInfo)
     }
 }
